# api/swagger.yaml
basePath: /api
definitions:
...
  portainer.Authorizations:
    additionalProperties:
      type: boolean
    type: object
...
  portainer.EndpointAuthorizations:
    additionalProperties:
      $ref: '#/definitions/portainer.Authorizations'
    type: object
...
  portainer.User:
    properties:
      EndpointAuthorizations:
        $ref: '#/definitions/portainer.EndpointAuthorizations'
      Id:
        description: User Identifier
        example: 1
        type: integer
      Password:
        example: passwd
        type: string
      PortainerAuthorizations:
        $ref: '#/definitions/portainer.Authorizations'
        description: |-
          Deprecated fields
          Deprecated in DBVersion == 25
      Role:
        description: User role (1 for administrator account and 2 for regular account)
        example: 1
        type: integer
      Username:
        example: bob
        type: string
    type: object
...
  users.userCreatePayload:
    properties:
      password:
        example: cg9Wgky3
        type: string
      role:
        description: User role (1 for administrator account and 2 for regular account)
        enum:
          - 1
          - 2
        example: 2
        type: integer
      username:
        example: bob
        type: string
    required:
      - password
      - role
      - username
    type: object
...
info:
  contact:
    email: info@portainer.io
  description: |
    Portainer API is an HTTP API served by Portainer. It is used by the Portainer UI and everything you can do with the UI can be done using the HTTP API.
    Examples are available at https://documentation.portainer.io/api/api-examples/
    You can find out more about Portainer at [http://portainer.io](http://portainer.io) and get some support on [Slack](http://portainer.io/slack/).

    # Authentication

    Most of the API endpoints require to be authenticated as well as some level of authorization to be used.
    Portainer API uses JSON Web Token to manage authentication and thus requires you to provide a token in the **Authorization** header of each request
    with the **Bearer** authentication mechanism.

    Example:

    ```
    Bearer abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890abcdefghijklmnopqrstuvwxyzAB
    ```

    # Security

    Each API endpoint has an associated access policy, it is documented in the description of each endpoint.

    Different access policies are available:

    - Public access
    - Authenticated access
    - Restricted access
    - Administrator access

    ### Public access

    No authentication is required to access the endpoints with this access policy.

    ### Authenticated access

    Authentication is required to access the endpoints with this access policy.

...

...

...

...

...

...

...

...

...
paths:
...
  /users:
    get:
      description: |-
        List Portainer users.
        Non-administrator users will only be able to list other non-administrator user accounts.
        **Access policy**: restricted
      operationId: UserList
      produces:
        - application/json
      responses:
        '200':
          description: Success
          schema:
            items:
              $ref: '#/definitions/portainer.User'
            type: array
        '400':
          description: Invalid request
        '500':
          description: Server error
      security:
        - jwt: []
      summary: List users
      tags:
        - users
    post:
      consumes:
        - application/json
      description: |-
        Create a new Portainer user.
        Only team leaders and administrators can create users.
        Only administrators can create an administrator user account.
        **Access policy**: restricted
      operationId: UserCreate
      parameters:
        - description: User details
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/users.userCreatePayload'
      produces:
        - application/json
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/portainer.User'
        '400':
          description: Invalid request
        '403':
          description: Permission denied
        '409':
          description: User already exists
        '500':
          description: Server error
      security:
        - jwt: []
      summary: Create a new user
      tags:
        - users
...
schemes:
...
securityDefinitions:
  jwt:
    in: header
    name: Authorization
    type: apiKey
swagger: '2.0'
...